book_store_backend/scripts/requirements.txt
requests==2.31.0
psycopg[binary]==3.2.3
python-dotenv==1.0.0
